requirements.txt
aiohttp==3.3.2
astroid==1.6.5
async-timeout==3.0.0
attrs==18.1.0
autopep8==1.3.5
backcall==0.1.0
beautifulsoup4==4.4.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.3.0
Flask==1.0.2
Flask-Cors==3.0.4
gunicorn==19.8.1
idna==2.6
idna-ssl==1.0.1
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
multidict==4.3.1
newrelic==3.4.0.95
numpy==1.14.3
pandas==0.22.0
parso==0.2.1
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
pycodestyle==2.4.0
Pygments==2.2.0
pylint==1.9.2
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
rope==0.10.7
scikit-learn==0.19.1
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
yarl==1.2.6
